fix AutoDisconnectHandler from triggering on tab changes
This fixes an issue where the SegmentedControl stops working when switching from a tab with a SegmentedControl and later navigating to the tab again.

The AutoDisconnectHandler was triggering when leaving a tab, because only the tabbedPages current element was returned by the GetNavigationTree method. This is problematic as of the time of evaluation the tabbedPage.CurrentElement is already set to the new tab that is navigated to.

GetNavigationTree now returns all contained tabs, to not prematurely disconnect the handlers.